# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "InNetwork API",
    "version": "3.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/ingest": {
      "post": {
        "description": "Token-secured refresh trigger for the scheduled ingest cron. Disabled (404) until\nINNETWORK_ADMIN_TOKEN is set; the ingest runs in the background so the cron returns\npromptly. Not under /api/, so it isn't rate-limited.",
        "operationId": "admin_ingest_admin_ingest_post",
        "parameters": [
          {
            "description": "all | tic | medicare",
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "default": "all",
              "description": "all | tic | medicare",
              "title": "Source",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Ingest Admin Ingest Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Ingest"
      }
    },
...
  }
}